+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.9.1
+
+- Support globally set plaster template using command line option `--plaster TEMPLATE`,
+  and plaster as a set instruction.
+
 ## 0.9.0
 
 - Support reading excerpts from `*.excerpt.yaml` files when the new `--yaml` flag is used.